Converted to support .Net Framework 4.7.1 and .Net Standard 2.0
Minor changes in unit tests to use NUnit
