--- Frends.Community.FixedWidthFlatFile.Tests/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="MSTest.TestAdapter" version="1.1.18" targetFramework="net461" />
-  <package id="MSTest.TestFramework" version="1.1.18" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="10.0.1" targetFramework="net461" />
-</packages>
