+++ Frends.Community.FixedWidthFlatFile.sln
@@ -1,19 +1,17 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27004.2010
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29613.14
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.FixedWidthFlatFile", "Frends.Community.FixedWidthFlatFile\Frends.Community.FixedWidthFlatFile.csproj", "{0E8C5741-16BE-4314-98B2-3F6C868F40B7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Frends.Community.FixedWidthFlatFile", "Frends.Community.FixedWidthFlatFile\Frends.Community.FixedWidthFlatFile.csproj", "{35C305C0-8108-4A98-BB1D-AFE5C926239E}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{BA7059DC-EAC6-4504-AA01-B2199D0C0EB0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Frends.Community.FixedWidthFlatFile.Tests", "Frends.Community.FixedWidthFlatFile.Tests\Frends.Community.FixedWidthFlatFile.Tests.csproj", "{8CA92187-8E4F-4414-803B-EC899479022E}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{78F7F22E-6E20-4BCE-8362-0C558568B729}"
 	ProjectSection(SolutionItems) = preProject
-		nuspec\Frends.Community.FixedWidthFlatFile.nuspec = nuspec\Frends.Community.FixedWidthFlatFile.nuspec
-		Frends.Community.FixedWidthFlatFile\FrendsTaskMetadata.json = Frends.Community.FixedWidthFlatFile\FrendsTaskMetadata.json
-		LICENSE = LICENSE
+		.gitignore = .gitignore
 		README.md = README.md
 	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.FixedWidthFlatFile.Tests", "Frends.Community.FixedWidthFlatFile.Tests\Frends.Community.FixedWidthFlatFile.Tests.csproj", "{6BE4B832-6F0E-45F0-82EF-DB749DCFFACD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,19 +19,19 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0E8C5741-16BE-4314-98B2-3F6C868F40B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0E8C5741-16BE-4314-98B2-3F6C868F40B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0E8C5741-16BE-4314-98B2-3F6C868F40B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0E8C5741-16BE-4314-98B2-3F6C868F40B7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6BE4B832-6F0E-45F0-82EF-DB749DCFFACD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6BE4B832-6F0E-45F0-82EF-DB749DCFFACD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6BE4B832-6F0E-45F0-82EF-DB749DCFFACD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6BE4B832-6F0E-45F0-82EF-DB749DCFFACD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6B2762FA-3388-48A3-B9E7-E53FBB2401CD}
+		SolutionGuid = {55BC6629-85C9-48D8-8CA2-B0046AF1AF4B}
 	EndGlobalSection
 EndGlobal
