+++ Frends.Community.FixedWidthFlatFile/FrendsTaskMetadata.json
@@ -1,7 +1,7 @@
 ﻿{
-  "Tasks": [
-    {
-      "TaskMethod": "Frends.Community.FixedWidthFlatFile.FixedWidthFlatFileTask.Parse"
-    }
-  ]
+	"Tasks": [
+		{
+			"TaskMethod": "Frends.Community.FixedWidthFlatFile.FixedWidthFlatFileTask.Parse"
+		}
+	]
 }
