--- nuspec/Frends.Community.FixedWidthFlatFile.nuspec
@@ -1,23 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-    <metadata>
-        <id>Frends.Community.FixedWidthFlatFile</id>
-        <version>1.2.0</version>
-        <title>Frends Community Fixed Width Flat File</title>
-        <authors>HiQ Finland</authors>
-        <requireLicenseAcceptance>false</requireLicenseAcceptance>
-        <description>Task for parsing and creating fixed width FlatFiles</description>
-        <summary />
-        <dependencies>
-            <dependency id="Newtonsoft.Json" version="6.0.8" />
-        </dependencies>
-        <frameworkAssemblies>
-            <frameworkAssembly assemblyName="System" targetFramework="" />
-        </frameworkAssemblies>
-    </metadata>
-    <files>
-        <file src="..\Frends.Community.FixedWidthFlatFile\bin\release\Frends.Community.FixedWidthFlatFile.dll" target="lib\net452\Frends.Community.FixedWidthFlatFile.dll" />
-        <file src="..\Frends.Community.FixedWidthFlatFile\bin\release\Frends.Community.FixedWidthFlatFile.xml" target="lib\net452\Frends.Community.FixedWidthFlatFile.xml" />
-        <file src="..\Frends.Community.FixedWidthFlatFile\FrendsTaskMetadata.json" target="FrendsTaskMetadata.json"/>
-    </files>
-</package>
